+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3894,7 +3894,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M0FW5WVMFJQJEKYV8APB20QG"),
+  id: z.string().default("01M0G7NMJH568K90SXG4J68RET"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),